[IUO] Migrate install and EUS upgrade paths to Konflux IDMS
CNV operator builds migrated from Brew to Konflux, so the install and
EUS upgrade test paths need Konflux IDMS instead of the legacy
ICSP/IDMS generated via `oc adm catalog mirror`.

This change consolidates Konflux IDMS helpers (apply_konflux_idms,
idms_has_all_mirrors) into tests/install_upgrade_operators/utils.py,
removes the legacy ICSP fixtures from both install and EUS upgrade
conftest files, and replaces them with new Konflux-based fixtures:

- Install path: installed_konflux_idms (product_install/conftest.py)
- EUS upgrade path: eus_updated_konflux_idms (product_upgrade/conftest.py)
- Upgrade path: add missing client param to updated_konflux_idms

Follows up on PR #4061 which fixed the upgrade path only.
